Added support for name scrolling
Long program names now scroll on the screen.

Fixed some code commenting errors.

Application names from the desktop app come padded with extra spaces. The 'end of name' is assumed to be when there are 2 spaces back to back.
This will be changed when the desktop app has been updated.