Fixing random mute/unmute on certain Arduino's (Tested Arduino Nano USB-C)
Fixing unexpected behavior of certain Arduino (clone?) boards where volume would glitch random value's or would randomly mute/unmute when turning the encoder.

This update might also increasing serial reliability for most other Atmega 328P boards.

Debugging the Glitches on an Arduino Nano (USB-C) clone it turns out either upping or lowering the boad rate to certain (common) rate would fix the problem. Digging into why this is the case revealed in many online conversations with the Atmega 328P datasheet backing this up that a baud rate of 115200 seems to be unfavored for the 328P chip in common configurations seen by Arduino boards. This baud rate matches difficult with the arduino configured 328P chip that it uses all correction/recovery potential in order to establish a working connection. When hardware deviations are added on top (for example a cheap 16Mhz Crystal running slightly slower) this could just tip over the correction capability and would result in faulty communications.

Looking at the datasheet the closest speed to 115200 with a great clock speed alignmet would be 76800 for a wide range of clock speed (8Mhz,12Mhz,16Mhz,18Mhz,20Mhz).
So the use of a 76800 baud rate leaves much more room to correct for cheap and unstable hardware situations.